+++ tests/phpunit/tests/list-more-custom-field-names.php
@@ -53,6 +53,12 @@ class List_More_Custom_Field_Names_Test extends WP_UnitTestCase {
 		$this->assertEquals( 47, $this->get_postmeta_form_limit() );
 	}
 
+	public function test_c2c_list_more_custom_field_names_filter_second_argument() {
+		add_filter( 'c2c_list_more_custom_field_names', function ( $limit, $old_limit ) { return $old_limit + 4; }, 10, 2 );
+
+		$this->assertEquals( 34, $this->get_postmeta_form_limit() );
+	}
+
 	// Test this last since the constant can't be unset
 	public function test_constant_takes_precendence() {
 		$limit = 175;